New Modal component, HastusAreaModal using that
