+++ src/components/overlays/OverlayContainer.tsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import Alert from './Alert';
 import Confirm from './Confirm';
+import Modal from './Modal';
 
 const OverlayContainer = () => (
     <>
         <Alert />
         <Confirm />
+        <Modal />
     </>
 );
