+++ src/components/sidebar/nodeView/hastusAreaModal.scss
@@ -1,0 +1,26 @@
+@import '../../shared/styles/common.scss';
+@import '../../shared/styles/flexbox';
+@import '../../shared/styles/form';
+
+.hastusAreaModal {
+    .hastusAreaForm {
+        width: 400px;
+        min-height: 125px;
+        padding: 20px;
+
+        .header {
+            font-weight: bold;
+            font-size: 20px;
+            display: flex;
+            justify-content: center;
+            padding: 10px;
+        }
+    }
+    .bottomBarButtons {
+        display: flex;
+        padding: 20px;
+        border-top: 1px solid #dedede;
+        flex-direction: row;
+        justify-content: center;
+    }
+}
